Move the backend spawning into serve/backend.rs
